Role management : restrict if role assignrd. #3157 #7164

+++ dotnet/corecomponents/net.atos.daf.ct2.role/RoleManagement.cs
@@ -57,6 +57,10 @@ namespace net.atos.daf.ct2.role
         {
             try
             {
+                if (await roleRepository.IsRoleAssigned(roleid) > 0)
+                {
+                    return -1;
+                }
                 int RoleId= await roleRepository.DeleteRole(roleid, Accountid);
                 // auditlog.AddLogs(userId,userId,1,"Delete Role", RoleId > 0,"Role Management", "Role Deleted With Role Id " + RoleId.ToString());
                 return RoleId;
